Added ProductUiMapper.kt , enhanced models for make reviewModel non-nullable

+++ shopping/src/main/java/com/example/shopping/data/model/ReviewSummaryEntity.kt
@@ -1,6 +1,6 @@
 package com.example.shopping.data.model
 
 data class ReviewSummaryEntity(
-    val reviewAverage: Double,
+    val reviewAverage: Float,
     val reviewCount: Int
 )
