+++ shopping/src/main/java/com/example/shopping/domain/model/ReviewSummary.kt
@@ -1,6 +1,6 @@
 package com.example.shopping.domain.model
 
 data class ReviewSummary(
-    val reviewAverage: Double,
+    val reviewAverage: Float,
     val reviewCount: Int
 )
