+++ shopping/src/main/java/com/example/shopping/remote/model/ReviewSummaryRemote.kt
@@ -1,13 +1,10 @@
 package com.example.shopping.remote.model
 
-
 import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
 
 data class ReviewSummaryRemote(
     @field:Json(name = "reviewAverage")
-    val reviewAverage: Double?,
+    val reviewAverage: Float?,
     @field:Json(name = "reviewCount")
     val reviewCount: Int?
 )
